requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
arrow==1.2.3
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
blessed==1.20.0
brotlipy==0.7.0
certifi==2022.6.15
cffi==1.15.0
chardet==5.0.0
charset-normalizer==2.1.0
clang==5.0
click==8.1.3
cloudpickle==2.0.0
croniter==1.3.8
cryptography==36.0.0
cycler==0.11.0
dateutils==0.6.12
deepdiff==6.2.3
dnspython==2.3.0
docker-pycreds==0.4.0
email-validator==1.3.1
environs==9.5.0
fastapi==0.88.0
fonttools==4.32.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.9.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.3
inquirer==3.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.2
lightning-cloud==0.5.30
lxml==4.9.1
Markdown==3.3.7
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.16.0
matplotlib
mdurl==0.1.2
mkl-fft
mkl-random
mkl-service
multidict==6.0.4
networkx==2.8.4
nltk==3.7
#numpy==1.21.5
numpy
oauthlib==3.2.0
opacus==1.1.2
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.8.6
packaging==21.3
pandas==1.4.2
pathtools==0.1.2
Pillow==9.0.1
pip==21.2.4
progress==1.6
protobuf==4.22.0
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.5
pyDeprecate==0.3.2
Pygments==2.14.0
PyJWT==2.6.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-editor==1.0.4
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
readchar==4.0.3
regex==2022.6.2
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
sentry-sdk==1.15.0
setproctitle==1.3.2
setuptools==61.2.0
six==1.15.0
sklearn==0.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
starlette==0.22.0
starsessions==1.3.0
summary==0.2.0
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.1.0
tqdm==4.64.0
traitlets==5.9.0
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.20.0
# uvloop==0.17.0  # Optional - doesn't work on Windows, skip if installation fails
wandb==0.13.10
watchfiles==0.18.1
wcwidth==0.2.6
websocket-client==1.3.2
websockets==10.4
Werkzeug==2.1.2
wheel==0.37.1
wrapt==1.12.1
yarl==1.8.2
zipp==3.8.0
h5py
torchmetrics==0.8.0
lightning-bolts==0.6.0.post1
lightning-lite==1.8.0
lightning-utilities==0.3.0
pytorch-lightning==1.8.0
# minepy  # Optional - uses scikit-learn mutual information as alternative

